chore: remove components from their indivdual folders
