+++ jsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "baseUrl": "src",
     "paths": {
+      "@app/*": ["./src/*"],
       "components": ["./src/components"]
     }
   },
