+++ src/components/Header.js
@@ -1,19 +1,34 @@
-import styled from "styled-components";
 import { Link } from "react-router-dom";
+import styled from "styled-components";
+import CKI_Logo from "../assets/images/cki_logo.png";
 
-export const Container = styled.div`
+export default function Header() {
+  return (
+    <Container>
+      <Wrapper to="/">
+        <Logo src={CKI_Logo} alt="Circle K International Logo"/>
+        <Content>
+          <Heading>Circle K</Heading>
+          <Subheading>Cal Poly Pomona</Subheading>
+        </Content>
+      </Wrapper>
+    </Container>
+  );
+}
+
+const Container = styled.div`
   display: flex;
   min-width: 220px;
   height: 60px;
   margin-bottom: 10px;
 `;
 
-export const Wrapper = styled(Link)`
+const Wrapper = styled(Link)`
   display: flex;
   text-decoration: none;
 `;
 
-export const Content = styled.div`
+const Content = styled.div`
   flex-direction: column;
   justify-content: center;
   margin-left: 10px;
@@ -23,7 +38,7 @@ export const Content = styled.div`
   flex-wrap: wrap;
 `;
 
-export const Heading = styled.div`
+const Heading = styled.div`
   color: ${props => props.theme.hue.gold};
   font-size: 20px;
   font-weight: 300;
@@ -33,7 +48,7 @@ export const Heading = styled.div`
   margin: 0;
 `
 
-export const Subheading = styled.div`
+const Subheading = styled.div`
   color: ${props => props.theme.hue.blue};
   font-size: 20px;
   text-transform: lowercase;
@@ -42,8 +57,8 @@ export const Subheading = styled.div`
   margin: 0;
 `
 
-export const Logo = styled.img`
+const Logo = styled.img`
   display: block;
   max-height: 60px;
   max-width: 60px;
-`;
+`;
