--- src/components/Hero/index.js
@@ -1,7 +1,0 @@
-export function Hero() {
-  return (
-    <div>
-      hero
-    </div>
-  );
-}
