--- src/components/Tenants/styles.js
@@ -1,22 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-    justify-content: center;
-    gap: 100px;
-    margin-bottom: 20px;
-    display: flex;
-    flex-wrap: wrap;
-    width: 100%;
-    flex: 1;
-`;
-
-export const Content = styled.div`
-  display: flex;
-  max-width: 300px;
-  flex-direction: column;
-  align-items: center;
-`;
-
-export const Describe = styled.p`
-  text-align: center;
-`;
