--- src/components/index.js
@@ -1,12 +1,0 @@
-export * from "./Anchor";
-export * from "./Hero";
-export * from "./Navbar";
-export * from "./Header";
-export * from "./Media";
-export * from "./Avatar";
-export * from "./Footer";
-export * from "./Event";
-export * from "./About";
-export * from "./Events";
-export * from "./Dango";
-export * from "./Tenants";
