+++ src/index.js
@@ -19,7 +19,6 @@ import {
   Home,
   About,
   Resources,
-  Join
 } from "./pages";
 
 const root = document.getElementById("root");
@@ -39,7 +38,6 @@ ReactDOM.render(
             <Route path="/" element={<Home/>}/>
             <Route path="/about" element={<About/>}/>
             <Route path="/resources" element={<Resources/>}/>
-            <Route path="/join" element={<Join/>}/>
           </Routes>
         </BrowserRouter>
       </HelmetProvider>
