--- src/pages/Join/index.js
@@ -1,10 +1,0 @@
-import { Navbar } from "../../components";
-
-export function Join() {
-  return (
-    <div>
-      <Navbar/>
-      join
-    </div>
-  );
-}
