+++ src/pages/Resources/index.js
@@ -1,4 +1,4 @@
-import { Navbar } from "../../components";
+import Navbar from "components/Navbar";
 
 export function Resources() {
   return (
