+++ src/pages/index.js
@@ -1,4 +1,3 @@
 export * from "./Home";
 export * from "./About";
 export * from "./Resources";
-export * from "./Join";
